URL関係をconst/urls.jsに集約＆switchByEnvのarrow function化
